requirements.txt
-e .
hatchling==1.27.0
packaging==24.2
pathspec==0.12.1
pluggy==1.5.0
trove-classifiers==2024.10.21.16
Pillow==10.3.0
